executor: migrate test-infra to testify for aggregate_test.go
