+++ executor/aggregate_test.go
@@ -1578,44 +1578,3 @@ func TestIssue23314(t *testing.T) {
 	res := tk.MustQuery("select col1 from t1 group by col1")
 	res.Check(testkit.Rows("16:40:20.01"))
 }
-
-func TestAggInDisk(t *testing.T) {
-	store, clean := testkit.CreateMockStore(t)
-	defer clean()
-	tk := testkit.NewTestKit(t, store)
-	tk.MustExec("use test")
-	tk.MustExec("set tidb_hashagg_final_concurrency = 1;")
-	tk.MustExec("set tidb_hashagg_partial_concurrency = 1;")
-	tk.MustExec("set tidb_mem_quota_query = 4194304")
-	tk.MustExec("drop table if exists t1")
-	tk.MustExec("create table t(a int)")
-	sql := "insert into t values (0)"
-	for i := 1; i <= 200; i++ {
-		sql += fmt.Sprintf(",(%v)", i)
-	}
-	sql += ";"
-	tk.MustExec(sql)
-	rows := tk.MustQuery("desc analyze select /*+ HASH_AGG() */ avg(t1.a) from t t1 join t t2 group by t1.a, t2.a;").Rows()
-	for _, row := range rows {
-		length := len(row)
-		line := fmt.Sprintf("%v", row)
-		disk := fmt.Sprintf("%v", row[length-1])
-		if strings.Contains(line, "HashAgg") {
-			require.False(t, strings.Contains(disk, "0 Bytes"))
-			require.True(t, strings.Contains(disk, "MB") ||
-				strings.Contains(disk, "KB") ||
-				strings.Contains(disk, "Bytes"))
-		}
-	}
-
-	// Add code cover
-	// Test spill chunk. Add a line to avoid tmp spill chunk is always full.
-	tk.MustExec("insert into t values(0)")
-	tk.MustQuery("select sum(tt.b) from ( select /*+ HASH_AGG() */ avg(t1.a) as b from t t1 join t t2 group by t1.a, t2.a) as tt").Check(
-		testkit.Rows("4040100.0000"))
-	// Test no groupby and no data.
-	tk.MustExec("drop table t;")
-	tk.MustExec("create table t(c int, c1 int);")
-	tk.MustQuery("select /*+ HASH_AGG() */ count(c) from t;").Check(testkit.Rows("0"))
-	tk.MustQuery("select /*+ HASH_AGG() */ count(c) from t group by c1;").Check(testkit.Rows())
-}
